requirements.txt
discord.py
firebase-admin
aioboto3
python-dotenv
pytest
pytest-asyncio
Flask
asyncio